pyproject.toml
[tool.poetry]
name = "pinry"
version = "2.1.9"
description = "Pinry, a tiling image board system for people who want to save, tag, and share images, videos and webpages in an easy to skim through format. It's open-source and self-hosted."
authors = ["winkidney <winkidney@gmail.com>"]
license = "BSD 2-Clause"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
requests = "^2.31.0"  # Updated: fixes CVE-2023-32681
django = ">2.2.17,<3"
pillow = ">=10.0.0"  # Updated: fixes multiple CVEs including buffer overflow
markdown = "*"
django-filter = "==2.4.0"
coreapi = "^2.3.3"
psycopg2-binary = "2.9.9"
django-taggit = "1.3.0"
django-braces = "^1.15.0"
django-compressor = "^4.0"
mock = "^4.0.3"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
setuptools = ">=70.0.0"  # Required for distutils on Python 3.12+
pyyaml = ">=6.0.1"  # Fix build issues on Python 3.12
django-ratelimit = "^4.1.0"  # Rate limiting for auth endpoints


[tool.poetry.dev-dependencies]
flake8 = "*"
qrcode = "*"
django-extensions = "<4"  # Must be <4 for Django 2.2 compatibility
mkdocs = "*"
mkdocs-material = "*"
jinja2 = "==3.0.3"
black = ">=23.0.0"

[tool.poetry.group.dev.dependencies]
jinja2 = "3.0.3"
mkdocs-material = "8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

